Moves some files around and added the now url routes
